Revert "feature/DEV-326: se agregan reportes de evaluaciones"
